Added Faircon mode

+++ app/src/main/java/com/example/faircon/business/domain/model/Controller.kt
@@ -1,7 +1,7 @@
 package com.example.faircon.business.domain.model
 
 data class Controller(
-    val fanSpeed: Float,
+    val fanSpeed: Int,
     val temperature: Float,
     val tevVoltage: Float
 )
